add tests for delegations lockings
