+++ contracts/sfc/Staker.sol
@@ -853,13 +853,12 @@ contract Stakers is Ownable, StakersConstants {
         require(stakers[stakerID].status == OK_STATUS, "staker should be active");
         require(lockDuration >= 86400 * 14 && lockDuration <= 86400 * 365, "incorrect duration");
         uint256 endTime = block.timestamp.add(lockDuration);
-        require(lockedStakes[stakerID].endTime > endTime, "staker's locked will finish first");
+        require(lockedStakes[stakerID].endTime >= endTime, "staker's locking will finish first");
         require(lockedDelegations[delegator][stakerID].endTime < endTime, "already locked up");
-//        require(stakers[stakerID].paidUntilEpoch == currentSealedEpoch, "not all rewards claimed"); // for rewards burning
         LockedAmount memory lStake = LockedAmount(
             currentSealedEpoch.add(1),
             block.timestamp,
-            block.timestamp.add(lockDuration));
+            endTime);
         lockedDelegations[delegator][stakerID] = lStake;
     }
